CocoaPodsTest: Turn two test cases into functional ones
The analyzer has functional tests with expected result (YAML) files for
each package manager. So, cocoapods should also have that. Turn these
two tests into functional tests to have that covered, but also to remove
dependencies on the Podfile class in order to enable refactoring it
without breaking the tests.

Signed-off-by: Frank Viernau <frank.viernau@here.com>